refactor(card): extract shared useCardPrompts from useCardActions
Pull the delete-confirm alert and move-cards modal out of the
deck-editor-bound useCardActions into a useCardPrompts composable so any
card-action call site can reuse the same dialogs, copy, and sfx.
